Provide default file for PDF output
When the application is called with no number of cards specified,
it tries to write to stdout.  Which is fine except in the case of PDF
output.

We fixed this by adding a `default_destination` class method to
`CardWriter`, which usually returns `sys.stdout`.  But the
`PdfWriter` inherited class returns a file name instead.